refactor(groups): address PR review feedback
- Extract the group-switcher dropdown out of GroupsIndicator into its
  own ActiveGroupSwitcher component.
- Narrow resolveVotesForScope's groupMemberIds param to Set<string>
  only, since every real caller already passes a Set; drops the
  unused array-input overload and its dedicated test.
