+++ src/components/layout/AppHeader/GroupsIndicator.tsx
@@ -1,19 +1,12 @@
 import { Suspense } from "react";
 import { Link } from "@tanstack/react-router";
-import { UserPlus, Users, ChevronDown } from "lucide-react";
+import { UserPlus } from "lucide-react";
 import { Skeleton } from "@/components/ui/skeleton";
-import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import { useAuth } from "@/contexts/AuthContext";
 import { useActiveGroup } from "@/hooks/useActiveGroup";
-import { useSetActiveGroupMutation } from "@/api/groups/useSetActiveGroupMutation";
 import { cn } from "@/lib/utils";
 import { TooltipButton } from "./TooltipButton";
+import { ActiveGroupSwitcher } from "./ActiveGroupSwitcher";
 
 const groupsButtonClassName =
   "bg-transparent border-purple-400/50 text-purple-300 hover:bg-purple-600 hover:text-white hover:border-purple-600 transition-colors";
@@ -50,7 +43,6 @@ function GroupsIndicatorContent({
 }) {
   const { activeGroup, activeGroupId, groups, hasGroups } =
     useActiveGroup(userId);
-  const setActiveGroupMutation = useSetActiveGroupMutation();
 
   if (!hasGroups) {
     return (
@@ -70,58 +62,14 @@ function GroupsIndicatorContent({
     );
   }
 
-  function handleSelect(groupId: string | null) {
-    if (groupId === (activeGroupId ?? null)) {
-      return;
-    }
-    setActiveGroupMutation.mutate({ userId, groupId });
-  }
-
   return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant="outline"
-          size={isMobile ? "sm" : "default"}
-          className={groupsButtonClassName}
-          aria-label={
-            isMobile
-              ? `Active group: ${activeGroup?.name || "Everyone"}`
-              : undefined
-          }
-        >
-          <Users className="h-4 w-4" />
-          {!isMobile && (
-            <span className="ml-2 flex items-center gap-1">
-              {activeGroup?.name || "Everyone"}
-              <ChevronDown className="h-3 w-3" />
-            </span>
-          )}
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent className="bg-gray-800 border-purple-400/30">
-        <DropdownMenuItem
-          onClick={() => handleSelect(null)}
-          className={cn(
-            "text-purple-100 hover:bg-purple-600/30",
-            !activeGroupId && "bg-purple-600/20",
-          )}
-        >
-          Everyone
-        </DropdownMenuItem>
-        {groups.map((group) => (
-          <DropdownMenuItem
-            key={group.id}
-            onClick={() => handleSelect(group.id)}
-            className={cn(
-              "text-purple-100 hover:bg-purple-600/30",
-              activeGroupId === group.id && "bg-purple-600/20",
-            )}
-          >
-            {group.name}
-          </DropdownMenuItem>
-        ))}
-      </DropdownMenuContent>
-    </DropdownMenu>
+    <ActiveGroupSwitcher
+      isMobile={isMobile}
+      userId={userId}
+      activeGroupId={activeGroupId}
+      activeGroup={activeGroup}
+      groups={groups}
+      className={groupsButtonClassName}
+    />
   );
 }
